Fix: Make locale persist throughout session SMDD
