--- app/(smdd)/smdd2024/layout.tsx
@@ -1,7 +1,0 @@
-export default function MainSmddLayout({
-  children
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
