+++ app/[lang]/(smdd)/smdd2024/components/case-studies.tsx
