+++ app/[lang]/(smdd)/smdd2024/components/hero.tsx
