+++ app/[lang]/(smdd)/smdd2024/components/language-picker.tsx
@@ -12,7 +12,7 @@ export default function LanguagePicker({ lang }: { lang: SupportedLanguage }) {
         <Link
           scroll={false}
           key={l}
-          href={`/smdd2024/${l}`}
+          href={`/${l}/smdd2024`}
           className={cn(
             "rounded-full px-6 py-2 text-sm group-[data-scroll='false']:text-white group-[data-scroll='true']:text-black",
             lang === l ? "bg-smdd-red text-white" : null
