+++ app/[lang]/(smdd)/smdd2024/components/projects.tsx
