+++ app/[lang]/(smdd)/smdd2024/components/trusted.tsx
