+++ app/[lang]/(smdd)/smdd2024/data.ts
